move user to top level package in model
