+++ app/models/communitarian/post_delay.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "user"
+require_relative "../user"
 
 module Communitarian
   class PostDelay
