+++ app/models/user.rb
